:recycle: refactor: 컨텐츠 목록 페이지 컴포넌트 분리
