+++ src/widgets/contents/ui/FloatingButton.tsx
@@ -1,0 +1,66 @@
+import { ChevronUp } from "lucide-react";
+import styled from "@emotion/styled";
+import { useEffect, useState } from "react";
+
+const FloatingButton = () => {
+  const [showScrollTop, setShowScrollTop] = useState(false);
+
+  // 스크롤 이벤트 핸들러
+  useEffect(() => {
+    const handleScroll = () => {
+      setShowScrollTop(window.scrollY > 300);
+    };
+
+    window.addEventListener("scroll", handleScroll);
+    return () => window.removeEventListener("scroll", handleScroll);
+  }, []);
+
+  const scrollToTop = () => {
+    window.scrollTo({
+      top: 0,
+      behavior: "smooth",
+    });
+  };
+
+  return (
+    <Button visible={showScrollTop} onClick={scrollToTop}>
+      <ChevronUp size={20} />
+    </Button>
+  );
+};
+
+export default FloatingButton;
+
+const Button = styled.button<{ visible: boolean }>`
+  position: fixed;
+  bottom: 2rem;
+  right: 2rem;
+  width: 3rem;
+  height: 3rem;
+  background-color: ${({ theme }) => theme.colors.primary[500]};
+  color: white;
+  border: none;
+  border-radius: 50%;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  transition: all 0.3s ease;
+  opacity: ${(props) => (props.visible ? 1 : 0)};
+  visibility: ${(props) => (props.visible ? "visible" : "hidden")};
+  z-index: 1000;
+
+  &:hover {
+    background-color: ${({ theme }) => theme.colors.primary[600]};
+    transform: translateY(-2px);
+    box-shadow: 0 6px 16px rgba(0, 0, 0, 0.2);
+  }
+
+  @media (max-width: 640px) {
+    bottom: 1.5rem;
+    right: 1.5rem;
+    width: 2.5rem;
+    height: 2.5rem;
+  }
+`;
